add pages
* add header and footer fragments

* refactor catalog, game, login

* return old registration

* подключены ссылки,
сколлапсил футер, хедер

* удалил секцию с пикчей,
сколлапсил футер, исправил отображение в одну линию

* подключил ссылки,
сколлапсил футер, хедер

* исправил отображение кнопки логина/аккаунта

* подрубил ссылки, сколлапсил футер, хедер

* изменил кол-во отобрадемых элементов

* прикрутил футер, хедер
добавил стили

* удалил ненужный скрипт

* add adding game to cart

* add adding to cart, cart page

* fix error handling

* add update user page and feature

* add create and update game

---------

+++ src/main/java/ru/aston/gamerent/service/AccountService.java
@@ -14,5 +14,5 @@ public interface AccountService {
 
     List<Account> findByGameId(Long gameId);
 
-    int numberOfAvailableAccounts(Long gameId);
+    long numberOfAvailableAccounts(Long gameId);
 }
